cvae/requirements.txt
mlx>=0.0.8
mlx-data
numpy
Pillow
